backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.24.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

# AI/ML - Core
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.3.1

# AI/ML - LLM Providers
google-generativeai==0.3.2
anthropic==0.8.1
openai==1.7.2

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Speech-to-Text
openai-whisper==20231117
ffmpeg-python==0.2.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Vector DB
chromadb==0.4.22

# YouTube
yt-dlp==2023.12.30

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2
python-slugify==8.0.1
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
